Created Item class and driver

+++ ClassDocs/A11/A11Labs/src/CollegeStudent.java
@@ -12,7 +12,7 @@ public class CollegeStudent extends Student {
     public String getMajor(){
         return this.major;
     }
-    public double getYear(){
+    public int getYear(){
         return this.year;
     }
     public void setMajor(String maj){
